add fusion rule tester for torch
