+++ nn_meter/builder/backend_meta/fusion_rule_tester/__init__.py
@@ -1,4 +1,3 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
-from .interface import generate_testcases, detect_fusion_rule
-from .generate_testcase import BaseTestCase
+from .interface import BaseTestCase, generate_testcases, detect_fusion_rule
