+++ nn_meter/builder/backend_meta/fusion_rule_tester/utils.py
@@ -26,15 +26,18 @@ __BUILTIN_OPERATORS__ = {
     "flatten": "Flatten",
     "split": "Split"
 }
+__BUILTIN_TESTCASES__ = {'MON'}
 
 __user_config_folder__ = os.path.expanduser('~/.nn_meter/config')
 __registry_cfg_filename__ = 'registry.yaml'
-__REG_OPERATORS__ = {}
+__REG_OPERATORS__, __REG_TESTCASES__ = {}, {}
 if os.path.isfile(os.path.join(__user_config_folder__, __registry_cfg_filename__)):
     with open(os.path.join(__user_config_folder__, __registry_cfg_filename__), 'r') as fp:
         registry_modules = yaml.load(fp, yaml.FullLoader)
     if "operators" in registry_modules:
         __REG_OPERATORS__ = registry_modules["operators"]
+    if "testcases" in registry_modules:
+        __REG_TESTCASES__ = registry_modules["testcases"]
 
 
 def get_operator_by_name(operator_name, input_shape, config = None):
@@ -62,9 +65,37 @@ def get_operator_by_name(operator_name, input_shape, config = None):
     return operator, output_shape, op_is_two_inputs
 
 
-def generate_model_for_testcase(op1, op2, input_shape, config):
-    from .build_models import SingleOpModel, TwoOpModel
-    from nn_meter.builder.nn_generator.tf_networks.utils import get_inputs_by_shapes
+def get_special_testcases_by_name(testcase, implement=None):
+    if testcase in __BUILTIN_TESTCASES__:
+        assert implement != None
+        if implement == 'tensorflow':
+            from .build_tf_models import MultipleOutNodes
+            return MultipleOutNodes
+        elif implement == 'torch':
+            from .build_torch_models import MultipleOutNodes
+            return MultipleOutNodes
+    else:
+        try:
+            testcase_info = __REG_TESTCASES__[testcase]
+            sys.path.append(testcase_info["package_location"])
+            testcase_module_name = testcase_info["class_name"]
+            testcase_module = importlib.import_module(testcase_info["class_module"])
+            testcase_cls = getattr(testcase_module, testcase_module_name)
+            return testcase_cls
+        except:
+            raise KeyError(f'Unsupported test case: {testcase}.')
+
+
+def generate_models_for_testcase(op1, op2, input_shape, config, implement):
+    if implement == 'tensorflow':
+        from .build_tf_models import SingleOpModel, TwoOpModel
+        from nn_meter.builder.nn_generator.tf_networks.utils import get_inputs_by_shapes
+    elif implement == 'torch':
+        from .build_torch_models import SingleOpModel, TwoOpModel
+        from nn_meter.builder.nn_generator.torch_networks.utils import get_inputs_by_shapes
+    else:
+        raise NotImplementedError('You must choose one implementation of kernel from "tensorflow" or "pytorch"')
+
     layer1, op1_output_shape, op1_is_two_inputs = get_operator_by_name(op1, input_shape, config)
     layer2, _, op2_is_two_inputs = get_operator_by_name(op2, op1_output_shape, config)
 
@@ -83,5 +114,41 @@ def generate_model_for_testcase(op1, op2, input_shape, config):
     return op1_model, op2_model, block_model, op1_shapes, op2_shapes, block_shapes
 
 
+def generate_single_model(op, input_shape, config, implement):
+    if implement == 'tensorflow':
+        from .build_tf_models import SingleOpModel
+        from nn_meter.builder.nn_generator.tf_networks.utils import get_inputs_by_shapes
+    elif implement == 'torch':
+        from .build_torch_models import SingleOpModel
+        from nn_meter.builder.nn_generator.torch_networks.utils import get_inputs_by_shapes
+    else:
+        raise NotImplementedError('You must choose one implementation of kernel from "tensorflow" or "pytorch"')
+
+    layer, _, is_two_inputs = get_operator_by_name(op, input_shape, config)
+
+    model = SingleOpModel(layer)
+    shapes = [input_shape] * (1 + is_two_inputs)
+    model(get_inputs_by_shapes(shapes))
+
+    return model, shapes
+
+
+def save_model(model, model_path, implement):
+    if implement == 'tensorflow':
+        from tensorflow import keras
+        from nn_meter.builder.nn_generator.tf_networks.utils import get_tensor_by_shapes
+        model['model'](get_tensor_by_shapes(model['shapes']))
+        keras.models.save_model(model['model'], model_path)
+    elif implement == 'torch':
+        pass
+    else:
+        import pdb; pdb.set_trace()
+        raise NotImplementedError('You must choose one implementation of kernel from "tensorflow" or "pytorch"')
+
+
 def list_operators():
     return __BUILTIN_OPERATORS__ + ["* " + item for item in list(__REG_OPERATORS__.keys())]
+
+
+def list_testcases():
+    return __BUILTIN_TESTCASES__ + ["* " + item for item in list(__REG_TESTCASES__.keys())]
